+++ src/index.ts
@@ -16,6 +16,7 @@ export {
 } from './errors/TokenProviderErrors';
 export type {
   AuthorizationCodeProviderConfig,
+  CfPasscodeProviderConfig,
   ClientCredentialsProviderConfig,
   DeviceFlowProviderConfig,
   OidcBrowserProviderConfig,
@@ -29,6 +30,7 @@ export type {
 export {
   AuthorizationCodeProvider,
   BaseTokenProvider,
+  CfPasscodeProvider,
   ClientCredentialsProvider,
   DeviceFlowProvider,
   OidcBrowserProvider,